+++ src/main/java/ru/javawebinar/topjava/model/AbstractNamedEntity.java
@@ -4,7 +4,7 @@ public abstract class AbstractNamedEntity extends AbstractBaseEntity {
 
     protected String name;
 
-    protected AbstractNamedEntity(Integer id, String name) {
+    protected AbstractNamedEntity(Long id, String name) {
         super(id);
         this.name = name;
     }
