+++ src/main/java/ru/javawebinar/topjava/repository/UserRepository.java
@@ -8,10 +8,10 @@ public interface UserRepository {
     User save(User user);
 
     // false if not found
-    boolean delete(int id);
+    boolean delete(Long id);
 
     // null if not found
-    User get(int id);
+    User get(Long id);
 
     // null if not found
     User getByEmail(String email);
