+++ src/main/java/ru/javawebinar/topjava/service/UserServiceImpl.java
@@ -27,12 +27,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void delete(int id) throws NotFoundException {
+    public void delete(Long id) throws NotFoundException {
         checkNotFoundWithId(repository.delete(id), id);
     }
 
     @Override
-    public User get(int id) throws NotFoundException {
+    public User get(Long id) throws NotFoundException {
         return checkNotFoundWithId(repository.get(id), id);
     }
