+++ src/main/java/ru/javawebinar/topjava/web/user/AbstractUserController.java
@@ -22,7 +22,7 @@ public abstract class AbstractUserController {
         return service.getAll();
     }
 
-    public User get(int id) {
+    public User get(Long id) {
         log.info("get {}", id);
         return service.get(id);
     }
@@ -33,12 +33,12 @@ public abstract class AbstractUserController {
         return service.create(user);
     }
 
-    public void delete(int id) {
+    public void delete(Long id) {
         log.info("delete {}", id);
         service.delete(id);
     }
 
-    public void update(User user, int id) {
+    public void update(User user, Long id) {
         log.info("update {} with id={}", user, id);
         assureIdConsistent(user, id);
         service.update(user);
